segregate calls within utils so that dockerinit does not link to openssl
  for the use case of having alternate crypto uses and keeping
  dockerinit statically linked.

Docker-DCO-1.1-Signed-off-by: Vincent Batts <vbatts@redhat.com> (github: vbatts)
